dis-ambiguate std::sqrt(int) calls

+++ src/likelytest.cc
@@ -34,8 +34,8 @@ lk::Parameters const &errors, double prec) {
     double error(0);
     if(fmin->haveCovariance()) error = fmin->getErrors()[0];
     boost::format fmt("%d %.4f %.4f %.4f %.4f\n");
-    std::cout << fmt % methodId % std::log10(lk::lastMinEvalCount)
-        % (lk::lastMinGradCount ? std::log10(lk::lastMinGradCount) : 0.)
+    std::cout << fmt % methodId % std::log10((double)lk::lastMinEvalCount)
+        % (lk::lastMinGradCount ? std::log10((double)lk::lastMinGradCount) : 0.)
         % (fmin->getMinValue() > 0 ? -std::log10(fmin->getMinValue()) : 0) % error;
 }
